fixed race issue, switched to rb trees. BUG BELOW the race condition had to do with some of cores 1-3 not finishing writing their file by the time 0 did the final concatenation resulting in a large number of lost words.
the switch to trees decreased runtime by about 55% for serial and anywhere from
8 to 18% for parallel.

BUG
with the race condition fixed, the parallel and serial versions produce the
same output everytime for 128 or fewer files. however, the parallel program has
issues starting at 256 files that get doubly worse as the number of files
doubles. 128 / 16 < 10 while 256 / 16 > 10 and due to the business with
file naming, this is likely the cause of the problem. however, i wasn't able to
find it after a bit of searching.